refactor: merge latest code and add util package
